requirements.txt
pywin32
pyautogui
keyboard
requests
Pillow